Extract shared visitor pattern helper in feather transforms
Consolidated duplicated AST traversal boilerplate from
removeDuplicateEnumMembers, sanitizeEnumAssignments, and
removeDuplicateSemicolons into a shared visitFeatherAST helper.

The helper handles:
- Null/undefined guard
- Array traversal via Core.visitChildNodes
- Object type filtering
- Recursive child node visitation

This reduces ~58 lines of duplicated code across 3 call sites to a
single 18-line reusable helper in utils.ts.
