+++ .gitignore
@@ -88,6 +88,7 @@ node_modules/
 jspm_packages/
 .husky/_/
 reports/
+report/
 
 # Snowpack dependency directory (https://snowpack.dev/)
 web_modules/
